refactor: create custom hook and improve folder structure
