+++ client/src/components/Schedule/AddEventModal/index.ts
@@ -1,0 +1,1 @@
+export { default as AddEventModal } from './AddEventModal';
