+++ client/src/components/Schedule/EventDetails/EventDetails.tsx
@@ -6,9 +6,9 @@ import { CloseOutlined, EditOutlined } from '@ant-design/icons';
 import moment from 'moment-timezone';
 import css from 'styled-jsx/css';
 import { CourseEvent } from 'services/course';
-import { DEFAULT_COLORS } from './ScheduleSettings/scheduleSettingsHandlers';
+import { DEFAULT_COLORS } from '../constants';
 import { renderTagWithStyle, tagsRenderer } from 'components/Table';
-import { GithubUserLink } from '../GithubUserLink';
+import { GithubUserLink } from '../../GithubUserLink';
 import { TASK_TYPES_MAP } from 'data/taskTypes';
 
 const { Title, Text } = Typography;
