+++ client/src/components/Schedule/EventDetails/index.ts
@@ -1,0 +1,2 @@
+export { default as EventDetails } from './EventDetails';
+export { default as TaskDetails } from './TaskDetails';
