+++ client/src/components/Schedule/ListView/index.ts
@@ -1,0 +1,1 @@
+export { default as ListView } from './ListView';
