+++ client/src/components/Schedule/ScheduleSettings/index.ts
@@ -1,0 +1,1 @@
+export { default as ScheduleSettings } from './ScheduleSettings';
