+++ client/src/components/Schedule/TableView/EditableCell.tsx
@@ -3,7 +3,7 @@ import { Form, Input, DatePicker, TimePicker, Select } from 'antd';
 import { Rule } from 'antd/lib/form';
 import { UserSearch } from 'components/UserSearch';
 import { CourseEvent } from 'services/course';
-import { EVENT_TYPES, SPECIAL_ENTITY_TAGS, TASK_TYPES } from './model';
+import { EVENT_TYPES, SPECIAL_ENTITY_TAGS, TASK_TYPES } from '../model';
 import { UserService } from 'services/user';
 
 const { Option } = Select;
