+++ client/src/components/Schedule/TableView/index.ts
@@ -1,0 +1,1 @@
+export { default as TableView } from './TableView';
