+++ client/src/components/Table/renderers.tsx
@@ -8,7 +8,7 @@ import {
   GithubOutlined,
 } from '@ant-design/icons';
 import { Tag, Tooltip, Typography } from 'antd';
-import { getTagStyle } from '../Schedule/ScheduleSettings/scheduleSettingsHandlers';
+import { getTagStyle } from 'components/Schedule/utils';
 
 const { Text } = Typography;
