+++ client/src/pages/course/schedule.tsx
@@ -1,5 +1,6 @@
+import moment from 'moment-timezone';
 import { useState, useMemo } from 'react';
-import { useAsync, useLocalStorage } from 'react-use';
+import { useAsync } from 'react-use';
 import { Col, Row, Select, Tooltip, Button, Form, Upload, message } from 'antd';
 import { UploadFile } from 'antd/lib/upload/interface';
 import { RcFile } from 'antd/lib/upload';
@@ -16,59 +17,27 @@ import { withSession } from 'components/withSession';
 import { PageLayout } from 'components/PageLayout';
 import { TableView, CalendarView, ListView } from 'components/Schedule';
 import withCourseData from 'components/withCourseData';
-import { CourseEvent, CourseService, CourseTask, CourseTaskDetails } from 'services/course';
+import { CourseService } from 'services/course';
 import { CoursePageProps } from 'services/models';
 import { isCourseManager } from 'domain/user';
 import { TIMEZONES } from '../../configs/timezones';
-import { isMobileOnly } from 'mobile-device-detect';
-import { ViewMode, SPECIAL_TASK_TYPES, CONFIGURABLE_COLUMNS } from 'components/Schedule/model';
-import ScheduleSettings from 'components/Schedule/ScheduleSettings/ScheduleSettings';
-import { DEFAULT_COLORS } from 'components/Schedule/ScheduleSettings/scheduleSettingsHandlers';
-import ModalFormEntity from '../../components/Schedule/ModalFormEntity';
-import moment from 'moment-timezone';
-import { isUndefined } from 'lodash';
-import csv from 'csvtojson';
+import { ViewMode } from 'components/Schedule/model';
+import { ScheduleSettings } from 'components/Schedule/ScheduleSettings';
+import { AddEventModal } from 'components/Schedule/AddEventModal';
+import { tasksToEvents, parseFiles, uploadResults } from 'components/Schedule/utils';
+import useScheduleSettings from 'components/Schedule/useScheduleSettings';
 
 const { Option } = Select;
-const LOCAL_VIEW_MODE = 'scheduleViewMode';
-const TAG_COLORS = 'tagColors';
-const LOCAL_HIDE_OLD_EVENTS = 'scheduleHideOldEvents';
-const LOCAL_HIDE_DONE_TASKS = 'scheduleHideDoneTasks';
-const LIMIT_FOR_DONE_TASKS = 'scheduleLimitForDoneTask';
-const COLUMNS_SHOWN = 'scheduleColumnsShown';
-const EVENT_TYPES_SHOWN = 'eventTypesShown';
-const SPLITTED_BY_WEEK = 'scheduleSplitedByWeek';
 
 export function SchedulePage(props: CoursePageProps) {
   const [form] = Form.useForm();
   const [timeZone, setTimeZone] = useState(Intl.DateTimeFormat().resolvedOptions().timeZone);
-  const [scheduleViewMode, setScheduleViewMode] = useLocalStorage<string>(LOCAL_VIEW_MODE, getDefaultViewMode());
-  const [isOldEventsHidden, setOldEventsHidden] = useLocalStorage<boolean>(LOCAL_HIDE_OLD_EVENTS, false);
-  const [isDoneTasksHidden, setDoneTasksHidden] = useLocalStorage<boolean>(LOCAL_HIDE_DONE_TASKS, false);
-
   const [isModalOpen, setModalOpen] = useState(false);
   const [editableRecord, setEditableRecord] = useState(null);
   const [fileList, setFileList] = useState<RcFile[]>([]);
   const courseService = useMemo(() => new CourseService(props.course.id), [props.course.id]);
   const isAdmin = useMemo(() => isCourseManager(props.session, props.course.id), [props.session, props.course.id]);
-
-  const [splittedByWeek = false, setSplittedByWeek] = useLocalStorage<boolean>(SPLITTED_BY_WEEK, false);
-  const [eventTypesHidden = [], setEventTypesHidden] = useLocalStorage<string[]>(EVENT_TYPES_SHOWN, []);
-  const [columnsShown = [], setColumnsShown] = useLocalStorage<string[]>(COLUMNS_SHOWN, CONFIGURABLE_COLUMNS);
-  const [eventTypeTagsColors = [], setEventTypeTagsColors] = useLocalStorage<object>(TAG_COLORS, DEFAULT_COLORS);
-  const [limitForDoneTask = 100, setLimitForDoneTask] = useLocalStorage<number>(LIMIT_FOR_DONE_TASKS, 100);
-  const settings = {
-    splittedByWeek,
-    setSplittedByWeek,
-    eventTypesHidden,
-    setEventTypesHidden,
-    columnsShown,
-    setColumnsShown,
-    eventTypeTagsColors,
-    setEventTypeTagsColors,
-    limitForDoneTask,
-    setLimitForDoneTask,
-  };
+  const settings = useScheduleSettings();
 
   const loadData = async () => {
     const [courseEvents, courseTasks] = await Promise.all([
@@ -83,15 +52,15 @@ export function SchedulePage(props: CoursePageProps) {
 
   const filteredEvents = useMemo(() => {
     let filteredData = events.slice(0);
-    if (isOldEventsHidden) {
+    if (settings.arePassedEventsHidden) {
       const yesterday = moment.utc().subtract(1, 'day');
       filteredData = filteredData.filter(({ dateTime }) => moment(dateTime).isAfter(yesterday, 'day'));
     }
-    if (isDoneTasksHidden && limitForDoneTask !== undefined) {
-      filteredData = filteredData.filter(({ done }) => !done || done < limitForDoneTask);
+    if (settings.areDoneTasksHidden) {
+      filteredData = filteredData.filter(({ done }) => !done || done < settings.limitForDoneTask);
     }
     return filteredData;
-  }, [events, isOldEventsHidden, isDoneTasksHidden]);
+  }, [events, settings.arePassedEventsHidden, settings.areDoneTasksHidden, settings.limitForDoneTask]);
 
   const mapScheduleViewToComponent = {
     [ViewMode.TABLE]: TableView,
@@ -99,16 +68,10 @@ export function SchedulePage(props: CoursePageProps) {
     [ViewMode.CALENDAR]: CalendarView,
   };
 
-  const viewMode = scheduleViewMode as ViewMode;
-  const ScheduleView = mapScheduleViewToComponent[viewMode] || TableView;
+  const ScheduleView = mapScheduleViewToComponent[settings.viewMode] ?? TableView;
 
-  const toggleOldEvents = () => {
-    setOldEventsHidden(!isOldEventsHidden);
-  };
-
-  const toggleDoneTasks = () => {
-    setDoneTasksHidden(!isDoneTasksHidden);
-  };
+  const togglePassedEventsVisibility = () => settings.setArePassedEventsHidden(!settings.arePassedEventsHidden);
+  const toggleDoneTasksVisibility = () => settings.setAreDoneTasksHidden(!settings.areDoneTasksHidden);
 
   const closeModal = async () => {
     setEditableRecord(null);
@@ -142,21 +105,15 @@ export function SchedulePage(props: CoursePageProps) {
     }
   };
 
-  const onRemove = (_: UploadFile<any>) => {
-    setFileList([]);
-  };
+  const onRemove = (_: UploadFile<any>) => setFileList([]);
 
-  const beforeUpload = (file: RcFile) => {
-    setFileList([...fileList, file]);
-
-    return false;
-  };
+  const beforeUpload = (file: RcFile) => Boolean(setFileList([...fileList, file]));
 
   return (
     <PageLayout loading={loading} title="Schedule" githubId={props.session.githubId}>
       <Row justify="start" gutter={[16, 16]} style={{ marginBottom: '24px' }}>
         <Col>
-          <Select style={{ width: 100 }} defaultValue={scheduleViewMode} onChange={setScheduleViewMode}>
+          <Select style={{ width: 100 }} defaultValue={settings.viewMode} onChange={settings.setViewMode}>
             <Option value={ViewMode.TABLE}>Table</Option>
             <Option value={ViewMode.LIST}>List</Option>
             <Option value={ViewMode.CALENDAR}>Calendar</Option>
@@ -219,8 +176,8 @@ export function SchedulePage(props: CoursePageProps) {
           <Tooltip title="Hide old events" mouseEnterDelay={1}>
             <Button
               type="primary"
-              onClick={toggleOldEvents}
-              icon={isOldEventsHidden ? <EyeInvisibleOutlined /> : <EyeOutlined />}
+              onClick={togglePassedEventsVisibility}
+              icon={settings.arePassedEventsHidden ? <EyeInvisibleOutlined /> : <EyeOutlined />}
             />
           </Tooltip>
         </Col>
@@ -228,8 +185,8 @@ export function SchedulePage(props: CoursePageProps) {
           <Tooltip title="Hide done tasks" mouseEnterDelay={1}>
             <Button
               type="primary"
-              onClick={toggleDoneTasks}
-              icon={isDoneTasksHidden ? <BorderOutlined /> : <CheckSquareOutlined />}
+              onClick={toggleDoneTasksVisibility}
+              icon={settings.areDoneTasksHidden ? <BorderOutlined /> : <CheckSquareOutlined />}
             />
           </Tooltip>
         </Col>
@@ -243,14 +200,14 @@ export function SchedulePage(props: CoursePageProps) {
         isAdmin={isAdmin}
         courseId={props.course.id}
         refreshData={loadData}
-        tagColors={eventTypeTagsColors}
-        limitForDoneTask={limitForDoneTask}
+        tagColors={settings.tagColors}
+        limitForDoneTask={settings.limitForDoneTask}
         alias={props.course.alias}
-        columnsShown={columnsShown}
+        columnsShown={settings.columnsHidden}
         settings={settings}
       />
       {isModalOpen && (
-        <ModalFormEntity
+        <AddEventModal
           visible={isModalOpen}
           editableRecord={editableRecord}
           handleCancel={closeModal}
@@ -261,116 +218,5 @@ export function SchedulePage(props: CoursePageProps) {
     </PageLayout>
   );
 }
-
-const tasksToEvents = (tasks: CourseTaskDetails[]) => {
-  return tasks.reduce((acc: Array<CourseEvent>, task: CourseTaskDetails) => {
-    if (task.type !== SPECIAL_TASK_TYPES.test) {
-      acc.push(createCourseEventFromTask(task, task.type));
-    }
-    acc.push(
-      createCourseEventFromTask(
-        task,
-        task.type === SPECIAL_TASK_TYPES.test ? SPECIAL_TASK_TYPES.test : SPECIAL_TASK_TYPES.deadline,
-      ),
-    );
-    return acc;
-  }, []);
-};
-
-const createCourseEventFromTask = (task: CourseTaskDetails, type: string): CourseEvent => {
-  return {
-    id: task.id,
-    dateTime: (type === SPECIAL_TASK_TYPES.deadline ? task.studentEndDate : task.studentStartDate) || '',
-    event: {
-      type: type,
-      name: task.name,
-      descriptionUrl: task.descriptionUrl,
-      id: task.taskId,
-    },
-    organizer: {
-      githubId: task.taskOwner ? task.taskOwner.githubId : '',
-    },
-    isTask: true,
-    special: task.special,
-    duration: task.duration,
-    score: `${task.score ?? 0}/${task.maxScore}`,
-    done: task.score && task.maxScore ? Math.round((task.score / task.maxScore) * 100) : 0,
-  } as CourseEvent;
-};
-
-const getDefaultViewMode = () => {
-  const localView = localStorage.getItem(LOCAL_VIEW_MODE);
-
-  if (localView) {
-    return localView;
-  }
-
-  if (isMobileOnly) {
-    return ViewMode.LIST;
-  }
-
-  return ViewMode.TABLE;
-};
-
-const parseFiles = async (incomingFiles: any) => {
-  const files = incomingFiles.fileList;
-  const filesContent: string[] = await Promise.all(
-    files.map(
-      (file: any) =>
-        new Promise((res, rej) => {
-          const reader = new FileReader();
-          reader.readAsText(file.originFileObj, 'utf-8');
-          reader.onload = ({ target }) => res(target ? target.result : '');
-          reader.onerror = e => rej(e);
-        }),
-    ),
-  );
-
-  const entities = (await Promise.all(filesContent.map((content: string) => csv().fromString(content))))
-    .reduce((acc, cur) => acc.concat(cur), [])
-    .map((item: any) => {
-      if (isUndefined(item.entityType)) {
-        throw new Error('Incorrect data: CSV file should content the headers named "entityType"');
-      }
-      return {
-        entityType: item.entityType as string,
-        templateId: item.templateId as string,
-        id: item.id as string,
-        startDate: item.startDate as string,
-        endDate: item.endDate as string,
-        type: item.type as string,
-        special: item.special as string,
-        name: item.name as string,
-        descriptionUrl: item.descriptionUrl as string,
-        githubId: item.githubId as string,
-        place: item.place as string,
-        checker: item.checker as string,
-        pairsCount: item.pairsCount as string,
-      };
-    });
-
-  return entities;
-};
-
-const uploadResults = async (
-  courseService: CourseService,
-  data: {
-    entityType: string;
-    id: string;
-    startDate: string;
-    endDate: string;
-    type: string;
-    special: string;
-    name: string;
-    descriptionUrl: string;
-    githubId: string;
-    place: string;
-  }[],
-  timeZone: string,
-) => {
-  const resultsNewTasks = await courseService.postMultipleEntities(data as Partial<CourseEvent & CourseTask>, timeZone);
-
-  return resultsNewTasks;
-};
 
 export default withCourseData(withSession(SchedulePage));
